requirements.txt
asgiref==3.4.1
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.2.6
django-bootstrap-datepicker-plus==3.0.5
django-formtools==2.2
gunicorn==20.1.0
mysqlclient==2.0.3
Pillow==8.3.2
psycopg2-binary==2.9.1
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0
